Unit test for financial acl CRM_Member_BAO_MembershipType::getPermissionedMembershipTypes function
