+++ ext/financialacls/tests/phpunit/Civi/Financialacls/LineItemTest.php
@@ -1,10 +1,11 @@
 <?php
 
-use CRM_Financialacls_ExtensionUtil as E;
-use Civi\Test\HeadlessInterface;
-use Civi\Test\HookInterface;
-use Civi\Test\TransactionalInterface;
+namespace Civi\Financialacls;
+
 use Civi\Api4\PriceField;
+
+// I fought the Autoloader and the autoloader won.
+require_once 'BaseTestClass.php';
 
 /**
  * FIXME - Add test description.
@@ -20,25 +21,11 @@ use Civi\Api4\PriceField;
  *
  * @group headless
  */
-class LineItemTest extends \PHPUnit\Framework\TestCase implements HeadlessInterface, HookInterface, TransactionalInterface {
-
-  use Civi\Test\ContactTestTrait;
-  use Civi\Test\Api3TestTrait;
-
-  /**
-   * @return \Civi\Test\CiviEnvBuilder
-   * @throws \CRM_Extension_Exception_ParseException
-   */
-  public function setUpHeadless() {
-    // Civi\Test has many helpers, like install(), uninstall(), sql(), and sqlFile().
-    // See: https://docs.civicrm.org/dev/en/latest/testing/phpunit/#civitest
-    return \Civi\Test::headless()
-      ->installMe(__DIR__)
-      ->apply();
-  }
+class LineItemTest extends BaseTestClass {
 
   /**
    * Test api applies permissions on line item actions (delete & get).
+   *
    * @dataProvider versionThreeAndFour
    */
   public function testLineItemApiPermissions($version) {
@@ -51,13 +38,13 @@ class LineItemTest extends \PHPUnit\Framework\TestCase implements HeadlessInterf
         [
           'line_item' => [
             [
-              'financial_type_id' => CRM_Core_PseudoConstant::getKey('CRM_Contribute_BAO_Contribution', 'financial_type_id', 'Donation'),
+              'financial_type_id' => \CRM_Core_PseudoConstant::getKey('CRM_Contribute_BAO_Contribution', 'financial_type_id', 'Donation'),
               'line_total' => 40,
               'price_field_id' => $defaultPriceFieldID,
               'qty' => 1,
             ],
             [
-              'financial_type_id' => CRM_Core_PseudoConstant::getKey('CRM_Contribute_BAO_Contribution', 'financial_type_id', 'Member Dues'),
+              'financial_type_id' => \CRM_Core_PseudoConstant::getKey('CRM_Contribute_BAO_Contribution', 'financial_type_id', 'Member Dues'),
               'line_total' => 50,
               'price_field_id' => $defaultPriceFieldID,
               'qty' => 1,
@@ -68,18 +55,7 @@ class LineItemTest extends \PHPUnit\Framework\TestCase implements HeadlessInterf
     ]);
     $this->_apiversion = $version;
 
-    $this->setPermissions([
-      'access CiviCRM',
-      'access CiviContribute',
-      'edit contributions',
-      'delete in CiviContribute',
-      'view contributions of type Donation',
-      'delete contributions of type Donation',
-      'add contributions of type Donation',
-      'edit contributions of type Donation',
-    ]);
-    Civi::settings()->set('acl_financial_type', TRUE);
-    $this->createLoggedInUser();
+    $this->setupLoggedInUserWithLimitedFinancialTypeAccess();
 
     $lineItems = $this->callAPISuccess('LineItem', 'get', ['sequential' => TRUE])['values'];
     $this->assertCount(2, $lineItems);
@@ -103,19 +79,6 @@ class LineItemTest extends \PHPUnit\Framework\TestCase implements HeadlessInterf
 
     $this->callAPIFailure('LineItem', 'Create', ['id' => $line['id'], 'check_permissions' => TRUE, 'financial_type_id' => 'Event Fee']);
     $this->callAPISuccess('LineItem', 'Create', ['id' => $line['id'], 'check_permissions' => TRUE, 'financial_type_id' => 'Donation']);
-  }
-
-  /**
-   * Set ACL permissions, overwriting any existing ones.
-   *
-   * @param array $permissions
-   *   Array of permissions e.g ['access CiviCRM','access CiviContribute'],
-   */
-  protected function setPermissions($permissions) {
-    CRM_Core_Config::singleton()->userPermissionClass->permissions = $permissions;
-    if (isset(\Civi::$statics['CRM_Financial_BAO_FinancialType'])) {
-      unset(\Civi::$statics['CRM_Financial_BAO_FinancialType']);
-    }
   }
 
   /**
